@@ -11,8 +11,14 @@
 #include "fictrac_version.h"
 
 #include <string>
+#include <csignal>
+#include <memory>
 
-using std::string;
+using namespace std;
+
+/// Ctrl-c handling
+bool _active = true;
+void ctrlcHandler(int /*signum*/) { _active = false; }
 
 
 int main(int argc, char *argv[])
@@ -23,12 +29,13 @@ int main(int argc, char *argv[])
      PRINT("/// \tCONFIG_FN\tPath to input config file (defaults to config.txt).");
      PRINT("/// \tLOG_VERBOSITY\t[Optional] One of DBG, INF, WRN, ERR.");
      PRINT("///");
-     PRINT("/// Version: %2d.%02d (build date: %s)", FICTRAC_VERSION_MAJOR, FICTRAC_VERSION_MINOR, __DATE__);
+     PRINT("/// Version: %d.%d.%d (build date: %s)", FICTRAC_VERSION_MAJOR, FICTRAC_VERSION_MIDDLE, FICTRAC_VERSION_MINOR, __DATE__);
      PRINT("///\n");
 
 	/// Parse args.
 	string log_level = "info";
 	string config_fn = "config.txt";
+    bool do_stats = false;
 	for (int i = 1; i < argc; ++i) {
 		if ((string(argv[i]) == "--verbosity") || (string(argv[i]) == "-v")) {
 			if (++i < argc) {
@@ -38,8 +45,11 @@ int main(int argc, char *argv[])
                 LOG_ERR("-v/--verbosity requires one argument (debug < info (default) < warn < error)!");
 				return -1;
 			}
-		}
-		else {
+        }
+        else if (string(argv[i]) == "--stats") {
+            do_stats = true;
+        }
+        else {
             config_fn = argv[i];
 		}
 	}
@@ -47,8 +57,8 @@ int main(int argc, char *argv[])
     /// Set logging level.
     Logger::setVerbosity(log_level);
 
-	//// Catch cntl-c
-	//signal(SIGINT, TERMINATE);
+	// Catch cntl-c
+    signal(SIGINT, ctrlcHandler);
 
 	/// Set high priority (when run as SU).
     if (!SetProcessHighPriority()) {
@@ -57,18 +67,34 @@ int main(int argc, char *argv[])
         LOG("Set process priority to HIGH!");
     }
 
-    Trackball tracker(config_fn);
+    unique_ptr<Trackball> tracker = make_unique<Trackball>(config_fn);
 
     /// Now Trackball has spawned our worker threads, we set this thread to low priority.
     SetThreadNormalPriority();
 
-    // wait for tracking to finish
-    while (tracker.isActive()) { sleep(500); }
+    /// Wait for tracking to finish.
+    while (tracker->isActive()) {
+        if (!_active) {
+            tracker->terminate();
+        }
+        sleep(250);
+    }
 
-    //tracker.printState();
-    tracker.writeTemplate();
+    /// Save the eventual template to disk.
+    tracker->writeTemplate();
 
-    PRINT("\n\nHit ENTER to exit..");
-    getchar_clean();
+    /// If we're running in test mode, print some stats.
+    if (do_stats) {
+        tracker->dumpStats();
+    }
+
+    /// Try to force release of all objects.
+    tracker.reset();
+
+    /// Wait a bit before exiting...
+    sleep(250);
+
+    //PRINT("\n\nHit ENTER to exit..");
+    //getchar_clean();
     return 0;
 }
