@@ -13,6 +13,9 @@
 #include "FrameGrabber.h"
 #include "ConfigParser.h"
 
+// Stuff for shared memory output and control
+#include "SharedMemTransferData.h"
+
 /// OpenCV individual includes required by gcc?
 #include <opencv2/highgui.hpp>
 #include <opencv2/imgproc.hpp>  
@@ -33,17 +36,78 @@
 class Trackball
 {
 public:
+    /// Data.
+    struct DATA {
+        // trackball state
+        unsigned int cnt, seq;
+        CmPoint64f dr_roi, r_roi;
+        cv::Mat R_roi;
+        CmPoint64f dr_cam, r_cam;
+        cv::Mat R_cam;
+        CmPoint64f dr_lab, r_lab;
+        cv::Mat R_lab;
+        double ts, ms;
+
+        double velx, vely, step_mag, step_dir, intx, inty, heading, posx, posy;
+
+        // testing
+        double dist, ang_dist, step_avg, step_var, evals_avg;
+
+        // constructors
+        DATA()
+            : cnt(0), seq(0),
+            dr_roi(CmPoint64f(0, 0, 0)), r_roi(CmPoint64f(0, 0, 0)),
+            dr_cam(CmPoint64f(0, 0, 0)), r_cam(CmPoint64f(0, 0, 0)),
+            dr_lab(CmPoint64f(0, 0, 0)), r_lab(CmPoint64f(0, 0, 0)),
+            ts(-1),
+            velx(0), vely(0),
+            step_mag(0), step_dir(0),
+            intx(0), inty(0),
+            heading(0), posx(0), posy(0),
+            dist(0), ang_dist(0),
+            step_avg(0), step_var(0),
+            evals_avg(0)
+        {
+            R_roi = cv::Mat::eye(3, 3, CV_64F);
+            R_cam = cv::Mat::eye(3, 3, CV_64F);
+            R_lab = cv::Mat::eye(3, 3, CV_64F);
+        }
+
+        DATA(const DATA &d)
+            : cnt(d.cnt), seq(d.seq),
+            dr_roi(d.dr_roi), r_roi(d.r_roi),
+            dr_cam(d.dr_cam), r_cam(d.r_cam),
+            dr_lab(d.dr_lab), r_lab(d.r_lab),
+            ts(d.ts),
+            velx(d.velx), vely(d.vely),
+            step_mag(d.step_mag), step_dir(d.step_dir),
+            intx(d.intx), inty(d.inty),
+            heading(d.heading), posx(d.posx), posy(d.posy),
+            dist(d.dist), ang_dist(d.ang_dist),
+            step_avg(d.step_avg), step_var(d.step_var),
+            evals_avg(d.evals_avg)
+        {
+            R_roi = d.R_roi.clone();
+            R_cam = d.R_cam.clone();
+            R_lab = d.R_lab.clone();
+        }
+    };
+
+public:
     Trackball(std::string cfg_fn);
     ~Trackball();
 
     bool isActive() { return _active; }
-    void printState();
+    void terminate() { _kill = true; }
+    std::shared_ptr<Trackball::DATA> getState();
+    void dumpStats();
     bool writeTemplate(std::string fn = "");
 
 private:
     /// Worker function.
     void process();
 
+    void resetData();
     void reset();
     double testRotation(const double x[3]);
     virtual double objective(unsigned n, const double* x, double* grad) { return testRotation(x); }
@@ -55,6 +119,7 @@ private:
 private:
     /// Drawing
     struct DrawData {
+        unsigned int log_frame;
         cv::Mat src_frame, roi_frame, sphere_view, sphere_map;
         CmPoint64f dr_roi;
         cv::Mat R_roi;
@@ -79,14 +144,13 @@ private:
     std::unique_ptr<std::thread> _drawThread;
 
 private:
-
     ConfigParser _cfg;
 
     /// Camera models and remapping.
     CameraModelPtr _src_model, _roi_model, _sphere_model;
     RemapTransformPtr _cam_to_roi;
     cv::Mat _roi_to_cam_R, _cam_to_lab_R;
-    std::shared_ptr<double[]> _p1s_lut;
+    std::shared_ptr<std::vector<double>> _p1s_lut;
 
     /// Arrays.
     int _map_w, _map_h;
@@ -101,34 +165,27 @@ private:
     /// Optimisation.
     std::unique_ptr<Localiser> _localOpt, _globalOpt;
     double _error_thresh, _err;
-    bool _global_search;
+    bool _do_global_search;
     int _max_bad_frames;
     int _nevals;
 
     /// Program.
     bool _init, _reset, _clean_map;
 
-    /// Data.
-    unsigned int _cnt, _seq;
-    CmPoint64f _dr_roi, _r_roi;
-    cv::Mat _R_roi;
-    CmPoint64f _dr_cam, _r_cam;
-    cv::Mat _R_cam;
-    CmPoint64f _dr_lab, _r_lab;
-    cv::Mat _R_lab;
-    double _ts;
-
-    double _velx, _vely, _step_mag, _step_dir, _intx, _inty, _heading, _posx, _posy;
-    
-    // test data
-    double _dist, _ang_dist, _step_avg, _step_var, _evals_avg;
+    /// Data
+    DATA _data;
 
     /// Data i/o.
     std::string _base_fn;
     std::unique_ptr<FrameGrabber> _frameGrabber;
-    std::unique_ptr<Recorder> _log;
+    bool _do_sock_output, _do_com_output, _do_shmem_output;
+    std::unique_ptr<Recorder> _data_log, _data_sock, _data_com, _vid_frames;
 
     /// Thread stuff.
-    std::atomic_bool _active;
+    std::atomic_bool _active, _kill, _do_reset;
     std::unique_ptr<std::thread> _thread;
+
+
+    std::shared_ptr<SHMEMRegion<SHMEMTransferData>> _shmem_data;
+    std::shared_ptr<SHMEMRegion<SHMEMSignalData>> _shmem_signal;
 };
